// src/interface/branch.types.ts
export interface IBranch {
  plan_id?: string
  company_id: number
  branch_name: string
  branch_code: string
  email: string
  password: string
  address1: string
  address2: string
  city: string
  state: string
  country: string
  pincode: string
  GST_number: string
  pan_number: string
  registration_number: string
  branch_logo: string
}